removing problematic markdown filesand refactoring app to seperate redux store provider from the connect to store
