--- src/markdown/competencies/communication/listening/0pt.md
@@ -1,6 +1,0 @@
----
-category: 'Communication'
-proficiency: 'Listening'
-summary: 'Has patience to not interrupt, takes turns speaking, pays close attention to what is being shared, and maintains eye contact. Uses listening cues such as nodding to show engagement. Acknowledges concerns and different perspectives of others.'
-milestone: 0
----
