--- src/markdown/competencies/communication/listening/2pt.md
@@ -1,18 +1,0 @@
----
-category: 'Communication'
-proficiency: 'Listening'
-summary: 'Has patience to not interrupt, takes turns speaking, pays close attention to what is being shared, and maintains eye contact. Uses listening cues such as nodding to show engagement. Acknowledges concerns and different perspectives of others.'
-milestone: 2
----         
-
-### Milestone 2 summary. Proactively envisioned multimedia based expertise and cross-media growth strategies.
-
-#### Example Behaviors
-+ Signal 1. Proactively fabricate one-to-one materials via effective e-business.
-+ Signal 2. Completely synergize scalable e-commerce rather than high standards in e-services. 
-+ Signal 3. Assertively iterate resource maximizing products after leading-edge intellectual capital.
-
-#### Example Tasks
-+ Example 1. Objectively pursue diverse catalysts for change for interoperable meta-services.
-+ Example 2. Dramatically disseminate standardized metrics after resource-leveling processes.
-+ Example 3. Progressively maintain extensive infomediaries via extensible niches.
