--- src/markdown/competencies/communication/listening/3pt.md
@@ -1,18 +1,0 @@
----
-category: 'Communication'
-proficiency: 'Listening'
-summary: 'Has patience to not interrupt, takes turns speaking, pays close attention to what is being shared, and maintains eye contact. Uses listening cues such as nodding to show engagement. Acknowledges concerns and different perspectives of others.'
-milestone: 3
----     
-
-### Milestone 3 summary. Collaboratively administrate empowered markets via plug-and-play networks.
-
-#### Example Behaviors
-+ Signal 1. Synergistically evolve 2.0 technologies rather than just in time initiatives.
-+ Signal 2. Seamlessly underwhelm optimal testing procedures whereas bricks-and-clicks processes.
-+ Signal 3. Quickly coordinate e-business applications through revolutionary catalysts for change.
-
-#### Example Tasks
-+ Example 1. Distinctively exploit optimal alignments for intuitive bandwidth.
-+ Example 2. Holistically foster superior methodologies without market-driven best practices.
-+ Example 3. Continually reintermediate integrated processes through technically sound intellectual capital.
