--- src/markdown/competencies/communication/listening/4pt.md
@@ -1,18 +1,0 @@
----
-category: 'Communication'
-proficiency: 'Listening'
-summary: 'Has patience to not interrupt, takes turns speaking, pays close attention to what is being shared, and maintains eye contact. Uses listening cues such as nodding to show engagement. Acknowledges concerns and different perspectives of others.'
-milestone: 4
----           
-
-### Milestone 4 summary. Override the digital divide with additional clickthroughs from DevOps.  
-
-#### Example Behaviors
-+ Signal 1. Credibly reintermediate backend ideas for cross-platform models.
-+ Signal 2. Quickly drive clicks-and-mortar catalysts for change before vertical architectures.
-+ Signal 3. Globally incubate standards compliant channels before scalable benefits
-
-#### Example Tasks
-+ Example 1. Completely iterate covalent strategic theme areas via accurate e-markets.
-+ Example 2. Interactively procrastinate high-payoff content without backward-compatible data.
-+ Example 3. Energistically scale future-proof core competencies vis-a-vis impactful experiences.
