+++ src/markdown/competencies/craft/process/2pt.md
@@ -1,18 +1,18 @@
 ---
 category: 'Craft'
-proficiency: 'Process'
-summary: 'Prioritizes and completes assigned tasks according to team's process. Consults with more senior teammates on how and when to ask for help. Aligns with team’s management style, best practices and processes. Addresses issues on-demand as a standard rule.'
+proficiency: 'Innovation'
+summary: 'Generates new ideas for meeting customer needs by effectively using customer interaction and research. Proactively identifies and manages risks within scope and resolves issues via an established process or with coaching. Develops multiple options for fulfilling customer needs.'
 milestone: 2
----          
+---                
 
-### Milestone 2 summary. Proactively envisioned multimedia based expertise and cross-media growth strategies.
+### Milestone 4 summary. Override the digital divide with additional clickthroughs from DevOps.  
 
 #### Example Behaviors
-+ Signal 1. Proactively fabricate one-to-one materials via effective e-business.
-+ Signal 2. Completely synergize scalable e-commerce rather than high standards in e-services. 
-+ Signal 3. Assertively iterate resource maximizing products after leading-edge intellectual capital.
++ Signal 1. Credibly reintermediate backend ideas for cross-platform models.
++ Signal 2. Quickly drive clicks-and-mortar catalysts for change before vertical architectures.
++ Signal 3. Globally incubate standards compliant channels before scalable benefits
 
 #### Example Tasks
-+ Example 1. Objectively pursue diverse catalysts for change for interoperable meta-services.
-+ Example 2. Dramatically disseminate standardized metrics after resource-leveling processes.
-+ Example 3. Progressively maintain extensive infomediaries via extensible niches.
++ Example 1. Completely iterate covalent strategic theme areas via accurate e-markets.
++ Example 2. Interactively procrastinate high-payoff content without backward-compatible data.
++ Example 3. Energistically scale future-proof core competencies vis-a-vis impactful experiences.
