--- src/markdown/competencies/growth-mindset/adaptability/0pt.md
@@ -1,6 +1,0 @@
----
-category: 'Growth Mindset'
-proficiency: 'Adaptability'
-summary: 'Ability to cope with change positively and constructively. Actively seeks information to tests assumptions. Shifts approach in response to the demands of a changing situation by demonstrating resilience.'
-milestone: 0
----
