--- src/markdown/competencies/growth-mindset/adaptability/1pt.md
@@ -1,18 +1,0 @@
----
-category: 'Growth Mindset'
-proficiency: 'Adaptability'
-summary: 'Ability to cope with change positively and constructively. Actively seeks information to tests assumptions. Shifts approach in response to the demands of a changing situation by demonstrating resilience.'
-milestone: 1
----         
-
-### Milestone 1 summary. Collaboratively administrate turnkey channels whereas virtual e-tailers. 
-
-#### Example Behaviors
-+ Signal 1. Leverage agile frameworks to provide a robust synopsis for high level overviews.
-+ Signal 2. Iterative approaches to corporate strategy foster collaborative thinking to further the overall value proposition.
-+ Signal 3. Organically grow the holistic world view of disruptive innovation via workplace diversity and empowerment.
-
-#### Example Tasks
-+ Example 1. Bring to the table win-win survival strategies to ensure proactive domination.
-+ Example 2. Capitalize on low hanging fruit to identify a ballpark value added activity to beta test.
-+ Example 3. Dramatically engage top-line web services vis-a-vis cutting-edge deliverables.
