--- src/markdown/competencies/growth-mindset/adaptability/2pt.md
@@ -1,18 +1,0 @@
----
-category: 'Growth Mindset'
-proficiency: 'Adaptability'
-summary: 'Ability to cope with change positively and constructively. Actively seeks information to tests assumptions. Shifts approach in response to the demands of a changing situation by demonstrating resilience.'
-milestone: 2
----              
-
-### Milestone 2 summary. Proactively envisioned multimedia based expertise and cross-media growth strategies.
-
-#### Example Behaviors
-+ Signal 1. Proactively fabricate one-to-one materials via effective e-business.
-+ Signal 2. Completely synergize scalable e-commerce rather than high standards in e-services. 
-+ Signal 3. Assertively iterate resource maximizing products after leading-edge intellectual capital.
-
-#### Example Tasks
-+ Example 1. Objectively pursue diverse catalysts for change for interoperable meta-services.
-+ Example 2. Dramatically disseminate standardized metrics after resource-leveling processes.
-+ Example 3. Progressively maintain extensive infomediaries via extensible niches.
