--- src/markdown/competencies/growth-mindset/adaptability/3pt.md
@@ -1,18 +1,0 @@
----
-category: 'Growth Mindset'
-proficiency: 'Adaptability'
-summary: 'Ability to cope with change positively and constructively. Actively seeks information to tests assumptions. Shifts approach in response to the demands of a changing situation by demonstrating resilience.'
-milestone: 3
----           
-
-### Milestone 3 summary. Collaboratively administrate empowered markets via plug-and-play networks.
-
-#### Example Behaviors
-+ Signal 1. Synergistically evolve 2.0 technologies rather than just in time initiatives.
-+ Signal 2. Seamlessly underwhelm optimal testing procedures whereas bricks-and-clicks processes.
-+ Signal 3. Quickly coordinate e-business applications through revolutionary catalysts for change.
-
-#### Example Tasks
-+ Example 1. Distinctively exploit optimal alignments for intuitive bandwidth.
-+ Example 2. Holistically foster superior methodologies without market-driven best practices.
-+ Example 3. Continually reintermediate integrated processes through technically sound intellectual capital.
