--- src/markdown/competencies/growth-mindset/adaptability/4pt.md
@@ -1,18 +1,0 @@
----
-category: 'Growth Mindset'
-proficiency: 'Adaptability'
-summary: 'Ability to cope with change positively and constructively. Actively seeks information to tests assumptions. Shifts approach in response to the demands of a changing situation by demonstrating resilience.'
-milestone: 4
----                   
-
-### Milestone 4 summary. Override the digital divide with additional clickthroughs from DevOps.  
-
-#### Example Behaviors
-+ Signal 1. Credibly reintermediate backend ideas for cross-platform models.
-+ Signal 2. Quickly drive clicks-and-mortar catalysts for change before vertical architectures.
-+ Signal 3. Globally incubate standards compliant channels before scalable benefits
-
-#### Example Tasks
-+ Example 1. Completely iterate covalent strategic theme areas via accurate e-markets.
-+ Example 2. Interactively procrastinate high-payoff content without backward-compatible data.
-+ Example 3. Energistically scale future-proof core competencies vis-a-vis impactful experiences.
