--- src/markdown/competencies/growth-mindset/handling-ambiguity/0pt.md
@@ -1,6 +1,0 @@
----
-category: 'Growth Mindset'
-proficiency: 'Handling Ambiguity'
-summary: 'Functions effectively on limited information and not fully understanding the whole picture. Finds a way forward when there aren't clear next steps and works through open ended questions while maintaining a constructive attitude in times of stress. Can participate in creation of solid design with vague requirements. Can change strategies quickly when initial approach doesn’t work.'
-milestone: 0
----
