--- src/markdown/competencies/growth-mindset/handling-ambiguity/2pt.md
@@ -1,18 +1,0 @@
----
-category: 'Growth Mindset'
-proficiency: 'Handling Ambiguity'
-summary: 'Functions effectively on limited information and not fully understanding the whole picture. Finds a way forward when there aren't clear next steps and works through open ended questions while maintaining a constructive attitude in times of stress. Can participate in creation of solid design with vague requirements. Can change strategies quickly when initial approach doesn’t work.'
-milestone: 2
----                   
-
-### Milestone 2 summary. Proactively envisioned multimedia based expertise and cross-media growth strategies.
-
-#### Example Behaviors
-+ Signal 1. Proactively fabricate one-to-one materials via effective e-business.
-+ Signal 2. Completely synergize scalable e-commerce rather than high standards in e-services. 
-+ Signal 3. Assertively iterate resource maximizing products after leading-edge intellectual capital.
-
-#### Example Tasks
-+ Example 1. Objectively pursue diverse catalysts for change for interoperable meta-services.
-+ Example 2. Dramatically disseminate standardized metrics after resource-leveling processes.
-+ Example 3. Progressively maintain extensive infomediaries via extensible niches.
