--- src/markdown/competencies/growth-mindset/handling-ambiguity/3pt.md
@@ -1,18 +1,0 @@
----
-category: 'Growth Mindset'
-proficiency: 'Handling Ambiguity'
-summary: 'Functions effectively on limited information and not fully understanding the whole picture. Finds a way forward when there aren't clear next steps and works through open ended questions while maintaining a constructive attitude in times of stress. Can participate in creation of solid design with vague requirements. Can change strategies quickly when initial approach doesn’t work.'
-milestone: 3
----            
-
-### Milestone 3 summary. Collaboratively administrate empowered markets via plug-and-play networks.
-
-#### Example Behaviors
-+ Signal 1. Synergistically evolve 2.0 technologies rather than just in time initiatives.
-+ Signal 2. Seamlessly underwhelm optimal testing procedures whereas bricks-and-clicks processes.
-+ Signal 3. Quickly coordinate e-business applications through revolutionary catalysts for change.
-
-#### Example Tasks
-+ Example 1. Distinctively exploit optimal alignments for intuitive bandwidth.
-+ Example 2. Holistically foster superior methodologies without market-driven best practices.
-+ Example 3. Continually reintermediate integrated processes through technically sound intellectual capital.
