--- src/markdown/competencies/leadership/health-and-safety/0pt.md
@@ -1,6 +1,0 @@
----
-category: 'Leadership'
-proficiency: 'Health and Safety'
-summary: 'Has discipline to manage self-care through diet, exercise, relaxation, and fulfilling activities while keeping one's self and others from harm.'
-milestone: 0
----
