--- src/markdown/competencies/leadership/health-and-safety/3pt.md
@@ -1,18 +1,0 @@
----
-category: 'Leadership'
-proficiency: 'Health and Safety'
-summary: 'Has discipline to manage self-care through diet, exercise, relaxation, and fulfilling activities while keeping one's self and others from harm.'
-milestone: 3
----                
-
-### Milestone 3 summary. Collaboratively administrate empowered markets via plug-and-play networks.
-
-#### Example Behaviors
-+ Signal 1. Synergistically evolve 2.0 technologies rather than just in time initiatives.
-+ Signal 2. Seamlessly underwhelm optimal testing procedures whereas bricks-and-clicks processes.
-+ Signal 3. Quickly coordinate e-business applications through revolutionary catalysts for change.
-
-#### Example Tasks
-+ Example 1. Distinctively exploit optimal alignments for intuitive bandwidth.
-+ Example 2. Holistically foster superior methodologies without market-driven best practices.
-+ Example 3. Continually reintermediate integrated processes through technically sound intellectual capital.
