+++ src/pages/index.js
@@ -1,14 +1,19 @@
 import React from 'react'
+import { Provider } from 'react-redux'
 
 import Layout from '../components/layout/'
 import Login from '../components/login'
 import SEO from '../components/seo'
 
+import store from '../store/index.js'
+
 const IndexPage = () => (
-  <Layout>
-    <SEO title="Home" />
-    <Login />
-  </Layout>
+  <Provider store={store}>
+    <Layout>
+      <SEO title="Home" />
+      <Login />
+    </Layout>
+  </Provider>
 )
 
 export default IndexPage
